% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arborshiny-package.R
\docType{package}
\name{arborshiny-package}
\alias{arborShinyReport}
\alias{arborShinyReport-package}
\alias{arborshiny-package}
\title{arborshiny: Interactive Forest Inventory Report}
\description{
An interactive shiny report for 'arbor' QSM inventories.
}
\author{
\strong{Maintainer}: Jean-Romain Roussel \email{info@r-lidar.com}

Authors:
\itemize{
  \item Jean-Romain Roussel \email{info@r-lidar.com}
}

}
